apply permission for interface 2

+++ src/app/views/categories/personnels/personnel-list/personnel-list.component.ts
@@ -11,6 +11,7 @@ import { PagingParams } from 'src/app/shared/params/paging.param';
 
 import { PersonnelAddEditModalComponent } from '../modals/personnel-add-edit-modal/personnel-add-edit-modal.component';
 import { PersonnelViewDetailModalComponent } from '../modals/personnel-view-detail-modal/personnel-view-detail-modal.component';
+import { RoleService } from 'src/app/shared/services/role.service';
 
 @Component({
   selector: 'app-personnel-list',
@@ -40,6 +41,7 @@ export class PersonnelListComponent implements OnInit {
   constructor(
     private route: ActivatedRoute,
     private modalService: NzModalService,
+    private roleService: RoleService,
     private personnelService: PersonnelService,
     private notify: NotifyService) { }
 
@@ -67,10 +69,7 @@ export class PersonnelListComponent implements OnInit {
         this.loading = false;
         this.pagination = res.pagination;
         this.dataSet = res.result;
-      }, error => {
-        this.notify.error('Có lỗi xảy ra');
-        console.log('error getAllPagingPersonnel');
-      }, () => {
+
         if (this.dataSet.length === 0 && this.pagination.currentPage !== 1) {
           this.pagination.currentPage -= 1;
           this.loadData();
@@ -79,73 +78,87 @@ export class PersonnelListComponent implements OnInit {
   }
 
   addNew() {
-    const modal = this.modalService.create({
-      nzTitle: 'Thêm nhân sự',
-      nzContent: PersonnelAddEditModalComponent,
-      nzMaskClosable: false,
-      nzClosable: false,
-      nzWidth: 880,
-      nzComponentParams: {
-        personnel: {},
-        isAddNew: true
-      },
-      nzFooter: [
-        {
-          label: 'Hủy',
-          shape: 'default',
-          onClick: () => modal.destroy(),
-        },
-        {
-          label: 'Lưu',
-          type: 'primary',
-          onClick: (componentInstance) => {
-            componentInstance.saveChanges();
-          }
-        }
-      ]
-    });
+    this.roleService.checkPermission('NHAN_SU', 'Create')
+      .subscribe((response: boolean) => {
+        if (response) {
+          const modal = this.modalService.create({
+            nzTitle: 'Thêm nhân sự',
+            nzContent: PersonnelAddEditModalComponent,
+            nzMaskClosable: false,
+            nzClosable: false,
+            nzWidth: 880,
+            nzComponentParams: {
+              personnel: {},
+              isAddNew: true
+            },
+            nzFooter: [
+              {
+                label: 'Hủy',
+                shape: 'default',
+                onClick: () => modal.destroy(),
+              },
+              {
+                label: 'Lưu',
+                type: 'primary',
+                onClick: (componentInstance) => {
+                  componentInstance.saveChanges();
+                }
+              }
+            ]
+          });
 
-    modal.afterClose.subscribe((result: boolean) => {
-      if (result) {
-        this.loadData();
-      }
-    });
+          modal.afterClose.subscribe((result: boolean) => {
+            if (result) {
+              this.loadData();
+            }
+          });
+        } else {
+          this.notify.warning('Bạn không có quyền');
+        }
+      });
   }
 
   update(id: number) {
-    this.personnelService.getDetail(id).subscribe((personnel: Personnel) => {
-      const modal = this.modalService.create({
-        nzTitle: 'Sửa nhân sự',
-        nzContent: PersonnelAddEditModalComponent,
-        nzMaskClosable: false,
-        nzClosable: false,
-        nzWidth: 880,
-        nzComponentParams: {
-          personnel,
-          isAddNew: false
-        },
-        nzFooter: [
-          {
-            label: 'Hủy',
-            shape: 'default',
-            onClick: () => modal.destroy()
-          },
-          {
-            label: 'Lưu',
-            type: 'primary',
-            onClick: (componentInstance) => {
-              componentInstance.saveChanges();
-            }
-          }
-        ]
-      });
+    this.roleService.checkPermission('NHAN_SU', 'Update')
+      .subscribe((response: boolean) => {
+        if (response) {
+          this.personnelService.getDetail(id).subscribe((personnel: Personnel) => {
+            const modal = this.modalService.create({
+              nzTitle: 'Sửa nhân sự',
+              nzContent: PersonnelAddEditModalComponent,
+              nzMaskClosable: false,
+              nzClosable: false,
+              nzWidth: 880,
+              nzComponentParams: {
+                personnel,
+                isAddNew: false
+              },
+              nzFooter: [
+                {
+                  label: 'Hủy',
+                  shape: 'default',
+                  onClick: () => modal.destroy()
+                },
+                {
+                  label: 'Lưu',
+                  type: 'primary',
+                  onClick: (componentInstance) => {
+                    componentInstance.saveChanges();
+                  }
+                }
+              ]
+            });
 
-      modal.afterClose.subscribe((result: boolean) => {
-        if (result) {
-          this.loadData();
+            modal.afterClose.subscribe((result: boolean) => {
+              if (result) {
+                this.loadData();
+              }
+            });
+          });
+        } else {
+          this.notify.warning('Bạn không có quyền');
         }
       });
-    });
   }
 
   view(personnel: Personnel) {
@@ -169,19 +182,23 @@ export class PersonnelListComponent implements OnInit {
   }
 
   delete(id: number) {
-    this.notify.confirm('Bạn có chắc chắn muốn xóa không?', () => {
-      this.personnelService.delete(id).subscribe((res: boolean) => {
-        if (res) {
-          this.notify.success('Xóa thành công!');
-          this.loadData();
+    this.roleService.checkPermission('NHAN_SU', 'Delete')
+      .subscribe((response: boolean) => {
+        if (response) {
+          this.notify.confirm('Bạn có chắc chắn muốn xóa không?', () => {
+            this.personnelService.delete(id).subscribe((res: boolean) => {
+              if (res) {
+                this.notify.success('Xóa thành công!');
+                this.loadData();
+              } else {
+                this.notify.warning('Nhân sự đang được sử dụng. Không được xóa!');
+              }
+            });
+          });
         } else {
-          this.notify.warning('Nhân sự đang được sử dụng. Không được xóa!');
+          this.notify.warning('Bạn không có quyền');
         }
-      }, _ => {
-        this.notify.error('Có lỗi xảy ra');
-        console.log('error deletePersonnel');
       });
-    });
   }
 
   search(keyword: string) {
